+++ library/src/methods/brand/index.ts
@@ -1,0 +1,1 @@
+export { brand } from './brand.ts';