+++ library/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the library will be documented in this file.
 
 - Change input type of `mimeType` validation to `Blob`
 - Rename `useDefault` method to `withDefault` (issue #80)
+- Add `brand` method to support branded types (pull request #85)
 
 ## v0.11.1 (August 07, 2023)